Refactored Model Code
