+++ View/Admin/AdminScreensKivy.kv
@@ -13,6 +13,8 @@
             pos: self.pos
             size: self.size
 
+
+
     BoxLayout:
         orientation: 'vertical'
         size_hint: None, None
@@ -85,7 +87,6 @@
 
 <InventoryScreen>
     backButton: back
-
     grid: gridLayout
 
     canvas:
